Split up services in to proper applications of their own since they don't interact in any way
