+++ bookings/static/bookings/js/bookings.js
