+++ bookings/templates/bookings/bookings.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 
 {% block header_files %}
-    <link rel="stylesheet" href="{% static 'services/css/bookings.css' %}" />
+    <link rel="stylesheet" href="{% static 'bookings/css/bookings.css' %}" />
 {% endblock %}
 
 {% block content %}
@@ -71,6 +71,6 @@
 {% endblock %}
 
 {% block scripts %}
-    <script type="text/javascript" src="{% static  'services/js/pikaday.js' %}"></script>
-    <script type="text/javascript" src="{% static  'services/js/bookings.js' %}"></script>
+    <script type="text/javascript" src="{% static  'bookings/js/pikaday.js' %}"></script>
+    <script type="text/javascript" src="{% static  'bookings/js/bookings.js' %}"></script>
 {% endblock %}
