+++ bookings/templates/bookings/equipmentbooking.html
