+++ evaluations/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from evaluations.models import *
+
+# Register your models here.
+admin.site.register(CourseEvaluation)
