+++ evaluations/urls.py
@@ -1,0 +1,4 @@
+from django.conf.urls import patterns, include, url
+
+urlpatterns = patterns('evaluations.views',
+                       url(r'^$', 'evaluations', name='resources_evaluations'),)
