+++ exambank/templates/exambank/exambank.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 
 {% block header_files %}
-    <link rel="stylesheet" href="{% static 'services/css/exambank.css' %}" />
+    <link rel="stylesheet" href="{% static 'exambank/css/exambank.css' %}" />
 {% endblock %}
 
 {% block content %}
@@ -38,5 +38,5 @@
 {% endblock %}
 
 {% block scripts %}
-    <script type="text/javascript" src="{% static  'services/js/exambank.js' %}"></script>
+    <script type="text/javascript" src="{% static  'exambank/js/exambank.js' %}"></script>
 {% endblock %}
