+++ lockers/forms.py
@@ -1,2 +1,3 @@
 from django import forms
+from services.models import Locker
 from django.contrib.auth.models import User
