+++ lockers/static/services/js/lockers.js
