+++ lockers/templates/lockers/lockers.html
@@ -32,6 +32,6 @@
     <script type="text/javascript">
         var post_url = "{% url 'resources_lockers' %}";
     </script>
-    <script type="text/javascript" src="{% static 'services/js/lockers.js' %}"></script>
+    <script type="text/javascript" src="{% static 'lockers/js/lockers.js' %}"></script>
 
 {% endblock %}
