+++ lockers/urls.py
@@ -1,0 +1,4 @@
+from django.conf.urls import patterns, include, url
+
+urlpatterns = patterns('lockers.views',
+                       url(r'lockers/', 'lockers', name='resources_lockers'),)
