+++ mathsocwebsite/urls.py
@@ -4,6 +4,10 @@ from mathsocwebsite import views
 
 urlpatterns = patterns('',
     url(r'^admin/', include(admin.site.urls)),
+    url(r'^bookings/', include('bookings.urls')),
+    url(r'^evaluations/', include('evaluations.urls')),
+    url(r'^exambank/', include('exambank.urls')),
+    url(r'^lockers/', include('lockers.urls')),
     url(r'^resources/', include('services.urls')),
     url(r'^tangent/', include('tangent.urls')),
     url(r'^login/$', 'cas.views.login', name="login"),
