+++ services/admin.py
@@ -1,7 +1,1 @@
 from django.contrib import admin
-from services.models import *
-
-# Register your models here.
-admin.site.register(Exam)
-admin.site.register(CourseEvaluation)
-admin.site.register(Locker)
